pyproject.toml
[project]
name = "flask-proj"
version = "0.1.0"
description = ""
authors = [
    {name = "RithuGowda55",email = "druthigs2003@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aniso8601 (==10.0.0)",
    "attrs (==25.1.0)",
    "black (==25.1.0)",
    "blinker (==1.9.0)",
    "cfgv (==3.4.0)",
    "click (==8.1.8)",
    "distlib (==0.3.9)",
    "filelock (==3.17.0)",
    "flasgger (==0.9.7.1)",
    "flask (==3.1.0)",
    "flask-cors (==5.0.0)",
    "flask-restx (==1.3.0)",
    "identify (==2.6.7)",
    "importlib-resources (==6.5.2)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.5)",
    "jsonschema (==4.23.0)",
    "jsonschema-specifications (==2024.10.1)",
    "markupsafe (==3.0.2)",
    "mistune (==3.1.2)",
    "mypy-extensions (==1.0.0)",
    "nodeenv (==1.9.1)",
    "packaging (==24.2)",
    "pathspec (==0.12.1)",
    "platformdirs (==4.3.6)",
    "pre-commit (==4.1.0)",
    "psycopg2-binary (==2.9.10)",
    "pytz (==2025.1)",
    "pyyaml (==6.0.2)",
    "referencing (==0.36.2)",
    "rpds-py (==0.22.3)",
    "ruff (==0.9.7)",
    "six (==1.17.0)",
    "virtualenv (==20.29.2)",
    "werkzeug (==3.1.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
